+++ local-tests/tests/wrapped-keys/testEthereumSignMessageGeneratedKey.ts
