+++ local-tests/tests/wrapped-keys/testEthereumSignMessageWrappedKey.ts
