+++ local-tests/tests/wrapped-keys/testFailImportWrappedKeysWithExpiredSessionSig.ts
@@ -53,14 +53,13 @@ export const testFailImportWrappedKeysWithExpiredSessionSig = async (
     });
     console.log(res);
   } catch (e: any) {
-    console.log('❌ THIS IS EXPECTED: ', e);
-    console.log(e.message);
-
     if (
       e.message.includes(
         'There was a problem fetching from the database: Error: Invalid pkpSessionSig: Expired sessionSig'
       )
     ) {
+      console.log('✅ THIS IS EXPECTED: ', e);
+      console.log(e.message);
       console.log(
         '✅ testFailImportWrappedKeysWithExpiredSessionSig is expected to have an error'
       );
