+++ local-tests/tests/wrapped-keys/testFailImportWrappedKeysWithSamePrivateKey.ts
@@ -35,16 +35,19 @@ export const testFailImportWrappedKeysWithSamePrivateKey = async (
       keyType: 'K256',
     });
   } catch (e: any) {
-    console.log('❌ THIS IS EXPECTED: ', e);
-
-    // TODO!: The Lambda isn't updated for this, uncomment it once done
-    // if (e.message === 'There was a problem fetching from the database: Error: "The conditional request failed"') {
-    //   console.log(
-    //     '✅ testFailImportWrappedKeysWithSamePrivateKey is expected to have an error'
-    //   );
-    // } else {
-    //   throw e;
-    // }
+    if (
+      e.message.includes(
+        'There is already a wrapped key stored, either for the provided pkpAddress, or with the same dataToEncryptHash; a pkpAddress may only have 1 wrapped key, and a wrapped key may only be associated with a single pkpAddress.'
+      )
+    ) {
+      console.log('✅ THIS IS EXPECTED: ', e);
+      console.log(e.message);
+      console.log(
+        '✅ testFailImportWrappedKeysWithSamePrivateKey is expected to have an error'
+      );
+    } else {
+      throw e;
+    }
   }
 
   console.log('✅ testFailImportWrappedKeysWithSamePrivateKey');
