+++ local-tests/tests/wrapped-keys/testSignMessageWithSolanaEncryptedKey.ts
